Improve log readability and add image similarity metric
This update is centered around enhancing the readability of the logs by restructuring the log statements to avoid line breaks. In addition, an image similarity measurement feature is added, which calculates and displays the similarity percentage between two images. The calculation is based on the difference in hash values of compared images. This feature will facilitate better understanding of duplication matches.