Improve logging and exception handling in DedupEngine
This update enhances the logging mechanism within the DedupEngine by enabling tracking of the number of images present in the given path. It also refines the process of creating symbolic links by providing more detailed information on successful creation or failure. Additionally, an adjustment was made to the IntegrationTests to generate a filename using Guid for each initialization. This change is complemented with improved assertions in the tests.